optimize trigger creation
Changes to optimize how triggers are generated based on the mashling
definition json.

If mashlins json defines:
1. HTTP Trigger A with {port: 9000, path: /pets/:petId, method: GET}
2. HTTP Trigger B with {port: 9000, path: /pets/:petId, method: POST}

Then, the parser must not create two trigger objects, both listening on
port 9000. When the app runs, one of them will throw ‘port not
available’ error and the process will terminate.

The uniqueness of a trigger is defined in its setting (not the handler
settings in flogo-speak). So in the example above, the port value
becomes the test of uniqueness. If ’n’ number of triggers defined in
the mashling json have setting of {port:9000} then they are essentially
the same flogo trigger, albeit differently named within the scope of
the mashling json definition.

A trigger that gets picked up is determined by alphabetical order
(after the first trigger defined is added to the list).